chrome/content/mochitest/tests/test_filterListener.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Filter listeners tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">

  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    prepareFilterComponents(true);
    preparePrefs();

    FilterStorage.addSubscription(Subscription.fromURL("~fl~"));
    FilterStorage.addSubscription(Subscription.fromURL("~wl~"));
    FilterStorage.addSubscription(Subscription.fromURL("~il~"));
    FilterStorage.addSubscription(Subscription.fromURL("~eh~"));

    function checkKnownFilters(text, expected)
    {
      let result = {};
      for each (let matcher in [blacklistMatcher, whitelistMatcher])
      {
        let filters = [];
        for each (let filter in matcher.regexps)
        {
          if (filter.shortcut)
            is(filter.shortcut, null, "Shortcut of filter " + filter.text);
          filters.push(filter);
        }
        for (let shortcut in matcher.shortcutHash)
        {
          let filter = matcher.shortcutHash[shortcut];
          if (filter.shortcut != shortcut)
            is(filter.shortcut, shortcut, "Shortcut of filter " + filter.text);
          filters.push(filter);
        }
        result[matcher == blacklistMatcher ? "blacklist" : "whitelist"] = filters;
      }
      result.elemhide = getGlobalForObject(ElemHide).filters;

      function canonize(obj)
      {
        let result = [];
        for (let key in obj)
        {
          if (obj[key].length)
            result.push({key : key, value: obj[key].map(function(filter) {return filter.text;}).sort()});
        }
        result.sort(function(a, b)
        {
          if (a.key < b.key)
            return -1;
          else if (a.key > b.key)
            return 1;
          else
            return 0;
        });
        return result.map(function(entry) {return entry.key + ":\n" + entry.value.join("\n")}).join("\n");
      }

      is(canonize(result), canonize(expected), text);
    }

    let filter1 = Filter.fromText("filter1");
    let filter2 = Filter.fromText("@@filter2");
    let filter3 = Filter.fromText("#filter3");
    let filter4 = Filter.fromText("!filter4");

    FilterStorage.addFilter(filter1);
    checkKnownFilters("add filter1", {blacklist: [filter1]});
    FilterStorage.addFilter(filter2);
    checkKnownFilters("add filter2", {blacklist: [filter1], whitelist: [filter2]});
    FilterStorage.addFilter(filter3);
    checkKnownFilters("add filter3", {blacklist: [filter1], whitelist: [filter2], elemhide: [filter3]});
    FilterStorage.addFilter(filter4);
    checkKnownFilters("add filter4", {blacklist: [filter1], whitelist: [filter2], elemhide: [filter3]});

    FilterStorage.removeFilter(filter1);
    checkKnownFilters("remove filter1", {whitelist: [filter2], elemhide: [filter3]});
    filter2.disabled = true;
    FilterStorage.triggerFilterObservers("disable", [filter2]);
    checkKnownFilters("disable filter2", {elemhide: [filter3]});
    FilterStorage.removeFilter(filter2);
    checkKnownFilters("remove filter2", {elemhide: [filter3]});
    FilterStorage.removeFilter(filter4);
    checkKnownFilters("remove filter4", {elemhide: [filter3]});

    let subscription = Subscription.fromURL("http://test1/");
    subscription.filters = [filter1, filter2, filter3, filter4];

    FilterStorage.addSubscription(subscription);
    checkKnownFilters("add subscription with filter1, filter2, filter3, filter4", {blacklist: [filter1], elemhide: [filter3]});
    filter2.disabled = false;
    FilterStorage.triggerFilterObservers("enable", [filter2]);
    checkKnownFilters("enable filter2", {blacklist: [filter1], whitelist: [filter2], elemhide: [filter3]});
    FilterStorage.updateSubscriptionFilters(subscription, [filter4]);
    checkKnownFilters("change subscription filters to filter4", {elemhide: [filter3]});
    FilterStorage.removeFilter(filter3);
    checkKnownFilters("remove filter3", {});
    FilterStorage.updateSubscriptionFilters(subscription, [filter1, filter2]);
    checkKnownFilters("change subscription filters to filter1, filter2", {blacklist: [filter1], whitelist: [filter2]});
    FilterStorage.addFilter(filter1);
    checkKnownFilters("add filter1", {blacklist: [filter1], whitelist: [filter2]});

    filter1.disabled = true;
    FilterStorage.triggerFilterObservers("disable", [filter1]);
    checkKnownFilters("disable filter1", {whitelist: [filter2]});
    filter2.disabled = true;
    FilterStorage.triggerFilterObservers("disable", [filter2]);
    checkKnownFilters("disable filter2", {});
    filter1.disabled = false;
    filter2.disabled = false;
    FilterStorage.triggerFilterObservers("enable", [filter1, filter2]);
    checkKnownFilters("enable filter1, filter2", {blacklist: [filter1], whitelist: [filter2]});

    subscription.disabled = true;
    FilterStorage.triggerSubscriptionObservers("disable", [subscription]);
    checkKnownFilters("disable subscription", {blacklist: [filter1]});
    FilterStorage.removeSubscription(subscription);
    checkKnownFilters("remove subscription", {blacklist: [filter1]});
    FilterStorage.addSubscription(subscription);
    checkKnownFilters("add subscription", {blacklist: [filter1]});
    subscription.disabled = false;
    FilterStorage.triggerSubscriptionObservers("enable", [subscription]);
    checkKnownFilters("enable subscription", {blacklist: [filter1], whitelist: [filter2]});

    subscription.disabled = true;
    FilterStorage.triggerSubscriptionObservers("disable", [subscription]);
    checkKnownFilters("disable subscription", {blacklist: [filter1]});
    FilterStorage.addFilter(filter2);
    checkKnownFilters("add filter2", {blacklist: [filter1], whitelist: [filter2]});
    FilterStorage.removeFilter(filter2);
    checkKnownFilters("remove filter2", {blacklist: [filter1]});
    subscription.disabled = false;
    FilterStorage.triggerSubscriptionObservers("enable", [subscription]);
    checkKnownFilters("enable subscription", {blacklist: [filter1], whitelist: [filter2]});

    let subscription2 = Subscription.fromURL("~fl~");
    subscription2.disabled = true;
    FilterStorage.triggerSubscriptionObservers("disable", [subscription2]);
    checkKnownFilters("disable blocking filters", {blacklist: [filter1], whitelist: [filter2]});
    FilterStorage.removeSubscription(subscription);
    checkKnownFilters("remove subscription", {});
    subscription2.disabled = false;
    FilterStorage.triggerSubscriptionObservers("enable", [subscription2]);
    checkKnownFilters("enable blocking filters", {blacklist: [filter1]});

    let subscription3 = Subscription.fromURL("~wl~");
    subscription3.disabled = true;
    FilterStorage.triggerSubscriptionObservers("disable", [subscription3]);
    checkKnownFilters("disable exception rules", {blacklist: [filter1]});
    FilterStorage.addFilter(filter2);
    checkKnownFilters("add filter2", {blacklist: [filter1]});
    subscription3.disabled = false;
    FilterStorage.triggerSubscriptionObservers("enable", [subscription3]);
    checkKnownFilters("enable exception rules", {blacklist: [filter1], whitelist: [filter2]});

    let subscription4 = Subscription.fromURL("http://test/");
    let filter5 = Filter.fromText("filter5");
    let filter6 = Filter.fromText("@@filter6");
    let filter7 = Filter.fromText("!filter7");
    FilterStorage.updateSubscriptionFilters(subscription4, [filter5, filter6, filter7]);
    checkKnownFilters("update subscription not in the list yet", {blacklist: [filter1], whitelist: [filter2]});

    FilterStorage.addSubscription(subscription4);
    checkKnownFilters("add subscription to the list", {blacklist: [filter1, filter5], whitelist: [filter2, filter6]});

    FilterStorage.updateSubscriptionFilters(subscription4, [filter5, filter2, filter7]);
    checkKnownFilters("update subscription while in the list", {blacklist: [filter1, filter5], whitelist: [filter2]});

    subscription3.disabled = true;
    FilterStorage.triggerSubscriptionObservers("disable", [subscription3]);
    checkKnownFilters("disable exception rules", {blacklist: [filter1, filter5], whitelist: [filter2]});

    FilterStorage.removeSubscription(subscription4);
    checkKnownFilters("remove subscription from the list", {blacklist: [filter1]});

    subscription3.disabled = false;
    FilterStorage.triggerSubscriptionObservers("enable", [subscription3]);
    checkKnownFilters("enable exception rules", {blacklist: [filter1], whitelist: [filter2]});
  </script>
  </pre>
</body>
</html>
